Rename markdown utility file
