+++ index.html
@@ -499,7 +499,7 @@
     </div>
 
     <script src="src/kanban.js"></script>
-    <script src="src/markdown-it.js"></script>
+    <script src="src/markdown.js"></script>
     <script src="src/github.js"></script>
     <script src="src/status-cards.js"></script>
 </body>
